+++ cpp/include/ucxx/request_am.h
@@ -11,8 +11,8 @@
 
 #include <ucp/api/ucp.h>
 
-#include <ucxx/constructors.h>
 #include <ucxx/delayed_submission.h>
+#include <ucxx/detail/constructors.h>
 #include <ucxx/request.h>
 #include <ucxx/request_am_builder.h>
 #include <ucxx/typedefs.h>
