+++ cpp/include/ucxx/request_endpoint_close.h
@@ -9,8 +9,8 @@
 
 #include <ucp/api/ucp.h>
 
-#include <ucxx/constructors.h>
 #include <ucxx/delayed_submission.h>
+#include <ucxx/detail/constructors.h>
 #include <ucxx/request.h>
 #include <ucxx/request_endpoint_close_builder.h>
 #include <ucxx/typedefs.h>
