+++ cpp/src/request_tag_builder.cpp
@@ -6,7 +6,7 @@
 #include <utility>
 #include <variant>
 
-#include <ucxx/constructors.h>
+#include <ucxx/detail/constructors.h>
 #include <ucxx/detail/register_inflight_request.h>
 #include <ucxx/endpoint.h>
 #include <ucxx/request.h>
